+++ src/main/java/com/team2813/vision/CameraConstants.java
@@ -1,0 +1,9 @@
+package com.team2813.vision;
+
+class CameraConstants {
+  public static final String LIMELIGHT_CAMERA_NAME = "limelight";
+
+  private CameraConstants() {
+    throw new AssertionError("Not instantiable");
+  }
+}
