+++ src/main/java/com/team2813/vision/LimelightPosePublisher.java
@@ -1,9 +1,10 @@
 package com.team2813.vision;
 
+import static com.team2813.vision.VisionNetworkTables.getTableForLimelight;
+
 import com.ctre.phoenix6.Utils;
 import com.team2813.lib2813.limelight.BotPoseEstimate;
 import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.networktables.StructTopic;
 import edu.wpi.first.units.Units;
@@ -12,7 +13,6 @@ import java.util.Optional;
 import java.util.function.Supplier;
 
 public final class LimelightPosePublisher {
-  static final String CAMERA_NAME = "limelight";
   private final TimestampedStructPublisher<Pose2d> publisher;
   private final double timestampOffset;
 
@@ -24,12 +24,8 @@ public final class LimelightPosePublisher {
     timestampOffset =
         clocks.fpgaTimestampSupplier().get() - clocks.currentTimestampSupplier().get();
     StructTopic<Pose2d> topic =
-        getNetworkTable(ntInstance).getStructTopic("poseEstimate", Pose2d.struct);
+        getTableForLimelight(ntInstance).getStructTopic("poseEstimate", Pose2d.struct);
     publisher = new TimestampedStructPublisher<>(topic, Pose2d.kZero, clocks.fpgaTimestampSupplier);
-  }
-
-  public static NetworkTable getNetworkTable(NetworkTableInstance ntInstance) {
-    return VisionUtil.getTableForCamera(ntInstance, CAMERA_NAME);
   }
 
   /**
