+++ src/main/java/com/team2813/vision/MultiPhotonPoseEstimator.java
@@ -1,6 +1,7 @@
 package com.team2813.vision;
 
-import static com.team2813.vision.VisionUtil.getTableForCamera;
+import static com.team2813.vision.CameraConstants.LIMELIGHT_CAMERA_NAME;
+import static com.team2813.vision.VisionNetworkTables.getTableForCamera;
 
 import edu.wpi.first.apriltag.AprilTagFieldLayout;
 import edu.wpi.first.math.geometry.Pose2d;
@@ -50,7 +51,7 @@ public class MultiPhotonPoseEstimator implements AutoCloseable {
     }
 
     private Builder addCamera(String name, Transform3d transform, Optional<String> description) {
-      if (name.equals(LimelightPosePublisher.CAMERA_NAME)) {
+      if (name.equals(LIMELIGHT_CAMERA_NAME)) {
         throw new IllegalArgumentException(String.format("Invalid camera name: '%s'", name));
       }
       if (cameraConfigs.put(name, new CameraConfig(transform, description)) != null) {
